% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_table.R
\name{inter_table}
\alias{inter_table}
\title{Make a table that shows the components of a Win Ratio Model}
\usage{
inter_table(
  model_name,
  tot_comp,
  endpoint_names = NULL,
  number_endpoints = NULL,
  text_size = 9,
  footnote_text = ""
)
}
\arguments{
\item{model_name}{A list output from a win.stat model, obtained by setting summary.print to false.}

\item{tot_comp}{A numeric, which is the total number of person to person comparisons. Calculated by multiplying the number of people in the treatment arm and the number of people in the control arm}

\item{endpoint_names}{A vector or character, which is the names of the endpoints in the model}

\item{number_endpoints}{A numeric, which is the number of endpoints in the model. Must be specified of endpoint_names is not specified}

\item{text_size}{A numeric, which is the size of the font for the text in the table. Default is 9.}

\item{footnote_text}{A character, which is the text that the footnote consists of. Default is no footnote.}
}
\value{
A flextable, which consists of the win ratio, net benefit, and win odds at each endpoint.
}
\description{
Make a table that shows the components of a Win Ratio Model
}
\examples{
data_binary <- WINS::data_binary
res_binary <- WINS::win.stat(data = data_binary, ep_type = "binary", priority = c(1:3),
stratum.weight = "unstratified", arm.name = c("A","B"), alpha=0.05,
digit = 3,pvalue = "two-sided", summary.print = FALSE)
#Get the total number of comparisons
tot_comps <- sum(data_binary$arm == "A") * sum(data_binary$arm == "B")
inter_table(res_binary, tot_comp = tot_comps, number_endpoints = 3)

}
