% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_ratio_functions.R
\name{nice_table}
\alias{nice_table}
\title{Make nice tables for Win Ratio model output}
\usage{
nice_table(model_name, text_size = 9, footnote_text = "")
}
\arguments{
\item{model_name}{A list output from a win.stat model, obtained by setting summary.print to false.}

\item{text_size}{A numeric, which is the size of the font for the text in the table. Default is 9.}

\item{footnote_text}{A character, which is the text that the footnote consists of. Default is no footnote.}
}
\value{
A flextable with the win ratio, net benefit, and win odds, along with their subsequent 95\% CIs and P-Values
}
\description{
Make nice tables for Win Ratio model output
}
\examples{
data_binary <- WINS::data_binary
res_binary <- WINS::win.stat(data = data_binary, ep_type = "binary", priority = c(1:3),
stratum.weight = "unstratified", arm.name = c("A","B"), alpha=0.05,
digit = 3,pvalue = "two-sided", summary.print = FALSE)
nice_table(res_binary,9,"test")

}
